move data processing out of inboxView and into use-inbox-data hook
